backend/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from dotenv import load_dotenv
from backend.routers import scan

load_dotenv()

app = FastAPI(
    title="ChimeraShield API",
    description="AI-powered cybersecurity analysis backend",
    version="0.1.0"
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(scan.router)


@app.get("/")
def root():
    return {"status": "ok", "service": "ChimeraShield API", "version": "0.1.0"}


@app.get("/health")
def health():
    return {"status": "healthy"}
